* make sure that map/reduce query cannot be the destination in pattern match * RaveDB-12074 - final batch of tests
