+++ test/Tests.Infrastructure/RavenTestBase.cs
@@ -10,6 +10,7 @@ using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using System.Threading;
 using System.Threading.Tasks;
+using FastTests.Graph;
 using Raven.Client;
 using Raven.Client.Documents;
 using Raven.Client.Documents.Indexes;
@@ -770,6 +771,172 @@ namespace FastTests
             {
                 if (_frozen)
                     throw new InvalidOperationException("Options are frozen and cannot be changed.");
+            }
+        }
+
+        protected void CreateSimpleData(IDocumentStore store)
+        {
+            using (var session = store.OpenSession())
+            {
+                var entityA = new Entity{ Id = "entity/1", Name = "A" };
+                var entityB = new Entity{ Id = "entity/2", Name = "B" };
+                var entityC = new Entity{ Id = "entity/3", Name = "C" };
+
+                session.Store(entityA);
+                session.Store(entityB);
+                session.Store(entityC);
+
+                entityA.References = entityB.Id;
+                entityB.References = entityC.Id;
+                entityC.References = entityA.Id;
+
+                session.SaveChanges();
+            }
+        }
+
+        protected void CreateDataWithMultipleEdgesOfTheSameType(IDocumentStore store)
+        {
+            using (var session = store.OpenSession())
+            {
+                var arava = new Dog { Name = "Arava" }; //dogs/1
+                var oscar = new Dog { Name = "Oscar" }; //dogs/2
+                var pheobe = new Dog { Name = "Pheobe" }; //dogs/3
+
+                session.Store(arava);
+                session.Store(oscar);
+                session.Store(pheobe);
+
+                //dogs/1 => dogs/2
+                arava.Likes = new[] { oscar.Id };
+                arava.Dislikes = new[] { pheobe.Id };
+
+                //dogs/2 => dogs/1,dogs/3 (cycle!)
+                oscar.Likes = new[] { oscar.Id, pheobe.Id };
+                oscar.Dislikes = new string[0];
+
+                //dogs/3 => dogs/2
+                pheobe.Likes = new[] { oscar.Id };
+                pheobe.Dislikes = new[] { arava.Id };
+
+                session.SaveChanges();
+            }
+        }
+
+        protected void CreateMoviesData(IDocumentStore store)
+        {
+            using (var session = store.OpenSession())
+            {
+                var scifi = new Genre
+                {
+                    Id = "genres/1",
+                    Name = "Sci-Fi"
+                };
+
+                var fantasy = new Genre
+                {
+                    Id = "genres/2",
+                    Name = "Fantasy"
+                };
+
+                var adventure = new Genre
+                {
+                    Id = "genres/3",
+                    Name = "Adventure"
+                };
+
+                session.Store(scifi);
+                session.Store(fantasy);
+                session.Store(adventure);
+
+                var starwars = new Movie
+                {
+                    Id = "movies/1",
+                    Name = "Star Wars Ep.1",
+                    Genres = new List<string>
+                    {
+                        "genres/1",
+                        "genres/2"
+                    }
+                };
+
+                var firefly = new Movie
+                {
+                    Id = "movies/2",
+                    Name = "Firefly Serenity",
+                    Genres = new List<string>
+                    {
+                        "genres/2",
+                        "genres/3"
+                    }
+                };
+
+                var indianaJones = new Movie
+                {
+                    Id = "movies/3",
+                    Name = "Indiana Jones and the Temple Of Doom",
+                    Genres = new List<string>
+                    {
+                        "genres/3"
+                    }
+                };
+
+                session.Store(starwars);
+                session.Store(firefly);
+                session.Store(indianaJones);
+
+                session.Store(new User
+                {
+                    Id = "users/1",
+                    Name = "Jack",
+                    HasRated = new List<User.Rating>
+                    {
+                        new User.Rating
+                        {
+                            Movie = "movies/1",
+                            Score = 5
+                        },
+                        new User.Rating
+                        {
+                            Movie = "movies/2",
+                            Score = 7
+                        }
+                    }
+                });
+
+                session.Store(new User
+                {
+                    Id = "users/2",
+                    Name = "Jill",
+                    HasRated = new List<User.Rating>
+                    {
+                        new User.Rating
+                        {
+                            Movie = "movies/2",
+                            Score = 7
+                        },
+                        new User.Rating
+                        {
+                            Movie = "movies/3",
+                            Score = 9
+                        }
+                    }
+                });
+
+                session.Store(new User
+                {
+                    Id = "users/3",
+                    Name = "Bob",
+                    HasRated = new List<User.Rating>
+                    {
+                        new User.Rating
+                        {
+                            Movie = "movies/3",
+                            Score = 5
+                        }
+                    }
+                });
+
+                session.SaveChanges();
             }
         }
     }
